requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
websockets==14.1

# LLM clients
google-generativeai==0.8.4
openai==1.59.6
httpx==0.28.1

# Configuration & serialization
pyyaml==6.0.2
pydantic==2.10.4
pydantic-settings==2.7.0

# Git operations
gitpython==3.1.44

# Scheduling & concurrency
apscheduler==3.11.0
anyio==4.8.0

# Utilities
python-dateutil==2.9.0
aiofiles==24.1.0
structlog==24.4.0

# Dev / code quality
yapf==0.43.0
pytest==8.3.4
pytest-asyncio==0.25.2
